added linux and mac support for monitor
