+++ server/classes/Honeytoken.ts
@@ -8,7 +8,6 @@ export abstract class Honeytoken implements I_Honeytoken {
   creationDate: Date;
   expirationDate: Date;
   grade: number;
-  notes: string;
 
   constructor(
     token_id: string,
@@ -16,7 +15,6 @@ export abstract class Honeytoken implements I_Honeytoken {
     type: HoneytokenType,
     expirationDate: Date,
     grade: number,
-    notes: string,
   ) {
     this.token_id = token_id;
     this.group_id = group_id;
@@ -25,7 +23,6 @@ export abstract class Honeytoken implements I_Honeytoken {
     // TODO: if expiration date is initial - add 30 days from today
     this.expirationDate = expirationDate;
     this.grade = grade;
-    this.notes = notes;
   }
 
   getTokenID(): string {
@@ -56,12 +53,7 @@ export abstract class Honeytoken implements I_Honeytoken {
     //TODO: implement db request through alerts
     throw new Error('Method not implemented.');
   }
-  getNotes(): string {
-    return this.notes!;
-  }
-  setNotes(notes: string): void {
-    this.notes = notes;
-  }
+
   getGrade(): number {
     return this.grade;
   }
