+++ server/interfaces/honeytoken.ts
@@ -8,7 +8,6 @@ export interface I_Honeytoken {
   getExpirationDate(): Date;
   isExpired(): boolean;
   isTriggered(): boolean;
-  getNotes(): string;
 }
 
 export interface I_HoneytokenManager {
